# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StreamPay API",
    "version": "2.0.0",
    "description": "REST API for the StreamPay payment-streaming platform on Stellar.\n\nAll error responses use the canonical envelope:\n```json\n{ \"error\": { \"code\": \"ERROR_CODE\", \"message\": \"Human-readable detail.\", \"request_id\": \"req_...\" } }\n```\n\nv1 stream paths (`/api/streams`) are deprecated and will be removed. Use `/api/v2/streams`."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
  "security": [
    { "bearerAuth": [] }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Obtain a token via POST /api/auth/wallet, then pass it as `Authorization: Bearer <token>`."
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message", "request_id"],
            "properties": {
              "code": {
                "type": "string",
                "description": "Machine-readable error code.",
                "example": "NOT_FOUND"
              },
              "message": {
                "type": "string",
                "description": "Human-readable description safe to display to clients.",
                "example": "The requested stream does not exist."
              },
              "request_id": {
                "type": "string",
                "description": "Opaque identifier for this request, useful for log correlation.",
                "example": "req_01HZ9ABCDEF"
              }
            }
          }
        }
      },
...
      "StreamV2": {
        "type": "object",
        "required": ["id", "recipient", "rate", "status", "allowed_actions", "created_at", "settlement"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique stream identifier.",
            "example": "stream_lf3k9z"
          },
          "recipient": {
            "type": "string",
            "description": "Stellar public key of the payment recipient.",
            "example": "GABC1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ234567890ABCDEFG"
          },
          "rate": {
            "type": "string",
            "description": "Human-readable payment rate.",
            "example": "120 XLM/month"
          },
          "status": {
            "type": "string",
            "enum": ["draft", "active", "paused", "ended"],
            "description": "Current lifecycle state of the stream."
          },
          "allowed_actions": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Actions the authenticated user may perform on this stream. Replaces v1 `actions`.",
            "example": ["pause", "stop"]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "ISO-8601 creation timestamp. Replaces v1 `createdAt`.",
            "example": "2024-01-15T10:00:00.000Z"
          },
          "settlement": {
            "oneOf": [
              {
                "type": "object",
                "required": ["settled_at", "amount", "currency"],
                "properties": {
                  "settled_at": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2024-02-01T12:00:00.000Z"
                  },
                  "amount": {
                    "type": "string",
                    "example": "360 XLM"
                  },
                  "currency": {
                    "type": "string",
                    "example": "XLM"
                  }
                }
              },
              { "type": "null" }
            ],
            "description": "Settlement details once the stream has been settled, otherwise `null`. Clients must handle `null` explicitly."
          }
        }
      },
...
    },
    "responses": {
      "400": {
        "description": "Bad request — invalid input.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
            "example": {
              "error": {
                "code": "BAD_REQUEST",
                "message": "Request body must include 'recipient' and 'rate'.",
                "request_id": "req_01HZ9ABCDEF"
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized — missing or invalid bearer token.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
            "example": {
              "error": {
                "code": "UNAUTHORIZED",
                "message": "Bearer token required.",
                "request_id": "req_01HZ9ABCDEF"
              }
            }
          }
        }
      },
...
      "404": {
        "description": "Not found — the requested resource does not exist.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
            "example": {
              "error": {
                "code": "STREAM_NOT_FOUND",
                "message": "Stream 'stream-abc12345' not found.",
                "request_id": "req_01HZ9ABCDEF"
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal server error.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
            "example": {
              "error": {
                "code": "INTERNAL_SERVER_ERROR",
                "message": "An unexpected error occurred.",
                "request_id": "req_01HZ9ABCDEF"
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/v2/streams/{id}/stop": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Stream identifier.",
          "schema": { "type": "string", "example": "stream_lf3k9z" }
        }
      ],
      "post": {
        "operationId": "stopStreamV2",
        "summary": "Stop a payment stream",
        "description": "Permanently stops a stream, transitioning it to `ended` status. Valid from `active` or `paused` states. This action is irreversible.",
        "tags": ["Streams"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Stream stopped successfully.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StreamV2" },
                "examples": {
                  "stopped-from-active": {
                    "summary": "Active stream stopped, now ended",
                    "value": {
                      "id": "stream_lf3k9z",
                      "recipient": "GABC1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ234567890ABCDEFG",
                      "rate": "120 XLM/month",
                      "status": "ended",
                      "allowed_actions": [],
                      "created_at": "2024-01-15T10:00:00.000Z",
                      "settlement": null
                    }
                  },
                  "stopped-from-paused": {
                    "summary": "Paused stream stopped, now ended",
                    "value": {
                      "id": "stream_m7n2p1",
                      "recipient": "GABC1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ234567890ABCDEFG",
                      "rate": "50 XLM/week",
                      "status": "ended",
                      "allowed_actions": [],
                      "created_at": "2024-01-10T08:00:00.000Z",
                      "settlement": null
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/401" },
          "404": { "$ref": "#/components/responses/404" },
          "409": {
            "description": "Conflict — stream is not in `active` or `paused` status.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
                "example": {
                  "error": {
                    "code": "INVALID_STATE_TRANSITION",
                    "message": "Cannot stop a stream that is not active or paused.",
                    "request_id": "req_01HZ9ABCDEF"
                  }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/500" }
        }
      }
    },
...
    "/api/webhooks/dlq": {
      "post": {
        "operationId": "receiveDlqWebhook",
        "summary": "Receive a dead-letter-queue webhook event",
        "description": "Accepts failed webhook events for reprocessing. No authentication required (secured at the network/gateway level).",
        "security": [],
        "tags": ["Webhooks"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Arbitrary webhook event payload.",
                "example": { "event": "payment.failed", "id": "evt_001" }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event accepted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["received"],
                  "properties": {
                    "received": { "type": "boolean", "example": true }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/400" },
          "500": { "$ref": "#/components/responses/500" }
        }
      }
    },
...
  },
...
}